Fix #18 -LabelNotFoundException in Song Prediction
Add a `getSongData()` method to `DatasetUtil` that only gets the dataset
instead of a `LearningDataset` object. Use that method in the
`predict()` methods of all classifiers. Also, use that method in
`getSongwiseDataset()` too for getting the dataset.

The `predict()` method in every classifier used `getSongwiseDataset()` to get
a `LearningDataset` object and then only passed the `dataset` property
to the `predict()` method of the actual Smile Classifier. So, when
`getSongwiseDataset()` tried to get the label data, it threw the
exception as the label array is undefined/unnecessary.

Fixes: #18