Close #19 Update FeatureType and Create Config
Add length attribute to FeatureType.
Add Config.java that stores the most of the configurations
we are interested in (as static attributes).
Fix spelling of 'SpectralVariability' (from 'SpectralVariablility')

Most of the configurations for the learning models can be done from
Config.java. Also, the features to use can be specified through the
array of `FeatureType` present in that class.